requirements.txt
mysql-connector-python~=9.3.0
psycopg2-binary~=2.9.9
pymssql ~= 2.3.6
oracledb ~= 3.4.2
protobuf >= 6.33.5